Merge upstream master into the Nagoya University Species Weighting Scheme
(SWS) branch and resolve conflicts in the collision code.

Master added a GASTALLY template parameter (for gas/gas collision tally
computes) to the collision routines, while the Nagoya branch added SWS
collision variants. Both dimensions are combined:

- src/collide.h, src/collide.cpp: the collisions() dispatch now selects
  over ambipolar x nearcp x ngas_tally x ngroups x sws. The four SWS
  routines gained the GASTALLY template parameter so gas tally computes
  work under SWS; the GASTALLY blocks are compile-time guarded, so
  non-tally SWS behavior is byte-for-byte unchanged.
- src/collide_vss.h/.cpp: kept master's perform_collision and the SWS
  perform_collision_SWS/test_collision_SWS/attempt_collision_SWS members.
  Repaired a silent line-level auto-merge that had grafted master's
  declaration onto the old perform_collision body.
- src/create_particles.cpp: took master's custom per-species insertion
  logic (species_var_flag/cummulative_custom); SWS np scaling preserved.
- src/fix_emit_surf.h: kept master's class restructuring and preserved
  the SWS cummulative_weighted member.

Verified with a serial build and by running the collide, chem, ambipolar,
and gas/collision/tally examples, plus SWS and SWS+gas-tally smoke tests.

+++ src/grid_surf.cpp
@@ -516,7 +516,7 @@ void Grid::surf2grid_surf_algorithm(int outflag)
 
     irregular = new Irregular(sparta);
     int nrecv1 = irregular->create_data_uniform(nsend,proclist1,1);
-    char *rbuf1 = (char *) memory->smalloc(nrecv1*nbytes_surf,"surf2grid:rbuf");
+    char *rbuf1 = (char *) memory->smalloc((bigint)nrecv1*nbytes_surf,"surf2grid:rbuf");
     irregular->exchange_uniform(sbuf1,nbytes_surf,rbuf1);
     delete irregular;
 
@@ -571,7 +571,7 @@ void Grid::surf2grid_surf_algorithm(int outflag)
 
     irregular = new Irregular(sparta);
     int nrecv2 = irregular->create_data_uniform(nsend,proclist2,1);
-    Send2 *rbuf2 = (Send2 *) memory->smalloc(nrecv2*sizeof(Send2),"surf2grid:rbuf2");
+    Send2 *rbuf2 = (Send2 *) memory->smalloc((bigint)nrecv2*sizeof(Send2),"surf2grid:rbuf2");
     irregular->exchange_uniform((char *) sbuf2,sizeof(Send2),(char *) rbuf2);
     delete irregular;
 
@@ -716,7 +716,7 @@ void Grid::surf2grid_surf_algorithm(int outflag)
 
     irregular = new Irregular(sparta);
     int nrecv3 = irregular->create_data_uniform(nsend,proclist3,1);
-    Send3 *rbuf3 = (Send3 *) memory->smalloc(nrecv3*sizeof(Send3),
+    Send3 *rbuf3 = (Send3 *) memory->smalloc((bigint)nrecv3*sizeof(Send3),
                                              "surf2grid:rbuf3");
     irregular->exchange_uniform((char *) sbuf3,sizeof(Send3),(char *) rbuf3);
     delete irregular;
